log4js logging, clean up utils
