+++ .dockerignore
@@ -2,3 +2,4 @@ dist
 node_modules
 yarn-error.log
 .git
+logs
