+++ .gitignore
@@ -5,6 +5,7 @@ dist
 .env
 data
 sounds
+logs
 
 assets/parrot*/
