+++ package.json
@@ -27,6 +27,7 @@
     "@mikro-orm/core": "^4.1.0",
     "@mikro-orm/migrations": "^4.1.1",
     "@mikro-orm/postgresql": "^4.1.1",
+    "@mikro-orm/sql-highlighter": "^1.0.1",
     "axios": "^0.20.0",
     "bufferutil": "^4.0.1",
     "canvas": "^2.6.1",
@@ -42,8 +43,10 @@
     "js-yaml": "^3.14.0",
     "libsodium-wrappers": "^0.7.8",
     "lodash": "^4.17.20",
+    "log4js": "^6.3.0",
     "lorem-ipsum": "^2.0.3",
     "mime-types": "^2.1.27",
+    "miniget": "^2.0.1",
     "moment": "^2.29.1",
     "music-metadata": "^7.4.0",
     "pg": "^8.4.1",
