+++ src/commands/games/liarsdice.ts
@@ -4,10 +4,9 @@ import _ = require('lodash');
 import * as randomstring from 'randomstring';
 
 import { Command } from '..';
-import { client } from '../..';
-import { Embed } from '../../embed';
+import { getLogger } from '../../providers';
 import { CmdArgs, DiceObject, GameReactionCollector, LiarsDiceGame } from '../../types';
-import { hasFlags, spliceFlag } from '../../util';
+import { Embed, hasFlags, spliceFlag } from '../../util';
 
 class Dice {
   value!: number;
@@ -53,12 +52,12 @@ export class CommandLiarsDice implements Command {
   docs = [
     {
       usage: 'liarsdice -c, --create [-d diceAmount=5] [-n diceSides=6]',
-      description: 'create dice game'
+      description: 'create dice game',
     },
     {
       usage: 'liarsdice -s, --start',
-      description: 'start dice game (game creator only)'
-    }
+      description: 'start dice game (game creator only)',
+    },
   ];
 
   makeGameCode(existing: string[]): string {
@@ -69,7 +68,7 @@ export class CommandLiarsDice implements Command {
         'ld-' +
         randomstring.generate({
           length: 4,
-          charset: '1234567890'
+          charset: '1234567890',
         });
     } while (existing.includes(code));
 
@@ -112,7 +111,7 @@ export class CommandLiarsDice implements Command {
           .setDescription('```yaml\n' + state + '\n```');
         return msg.channel.send(embed);
       } catch (err) {
-        console.log(err);
+        getLogger(msg.id).error(err);
         return msg.channel.send('got error: \n```\n' + err + '\n```');
       }
     }
@@ -175,7 +174,7 @@ export class CommandLiarsDice implements Command {
       timeLeft,
       playerTags,
       diceAmount,
-      diceSides
+      diceSides,
     });
 
     const embedMessage = await msg.channel.send(embed);
@@ -189,7 +188,7 @@ export class CommandLiarsDice implements Command {
           timeLeft: timeLeft -= 5,
           playerTags,
           diceAmount,
-          diceSides
+          diceSides,
         })
       );
     }, 5000);
@@ -240,11 +239,11 @@ export class CommandLiarsDice implements Command {
       players: {},
       metadata: { diceAmount, diceSides },
       creatorId: msg.author?.id as string,
-      reactionCollector: collector
+      reactionCollector: collector,
     };
 
     liarsDice[code].players[msg.author?.id as string] = {
-      dice: Dice.array(diceAmount, diceSides)
+      dice: Dice.array(diceAmount, diceSides),
     };
   }
 
@@ -253,7 +252,7 @@ export class CommandLiarsDice implements Command {
     timeLeft,
     playerTags,
     diceAmount,
-    diceSides
+    diceSides,
   }: {
     gameCode: string;
     timeLeft: number;
