+++ src/commands/general/help.ts
@@ -1,8 +1,8 @@
 import { Message } from 'discord.js';
 
 import { Command, CommandDocs, commands } from '..';
-import { Embed } from '../../embed';
 import { CmdArgs } from '../../types';
+import { Embed } from '../../util';
 
 export class CommandHelp implements Command {
   cmd = ['help', 'h'];
