+++ src/commands/moderation/unban.ts
@@ -1,7 +1,6 @@
 import { Message } from 'discord.js';
 
 import { Command } from '..';
-import { client } from '../..';
 import { CmdArgs } from '../../types';
 
 export class CommandUnban implements Command {
@@ -15,7 +14,7 @@ export class CommandUnban implements Command {
     if (!msg.guild?.members.resolve(msg.author?.id as string)?.hasPermission('BAN_MEMBERS'))
       return msg.channel.send('you are missing `BAN_MEMBERS` permission');
     if (args.length !== 1) return msg.channel.send('expected 1 arg');
-    if (!msg.guild?.members.resolve(client.user?.id as string)?.hasPermission('BAN_MEMBERS'))
+    if (!msg.guild?.me?.hasPermission('BAN_MEMBERS'))
       return msg.channel.send('bot is missing `BAN_MEMBERS` permission');
     try {
       await msg.guild.members.unban(args._[0].replace(/(<@!|>)/g, ''));
