--- src/hypixelapi.d.ts
@@ -1,546 +1,0 @@
-export interface BedwarsStats {
-  _items_purchased_bedwars: number;
-  activeBedDestroy: string;
-  activeDeathCry: string;
-  activeGlyph: string;
-  activeIslandTopper: string;
-  activeKillEffect: string;
-  activeKillMessages: string;
-  activeNPCSkin: string;
-  activeProjectileTrail: string;
-  activeSprays: string;
-  activeVictoryDance: string;
-  activeWoodType: string;
-  beds_broken_bedwars: number;
-  beds_lost_bedwars: number;
-  Bedwars_openedChests: number;
-  Bedwars_openedCommons: number;
-  Bedwars_openedEpics: number;
-  Bedwars_openedRares: number;
-  Bedwars_openedLegendaries: number;
-  castle__items_purchased_bedwars: number;
-  castle_beds_lost_bedwars: number;
-  castle_deaths_bedwars: number;
-  castle_diamond_resources_collected_bedwars: number;
-  castle_emerald_resources_collected_bedwars: number;
-  castle_entity_attack_deaths_bedwars: number;
-  castle_entity_attack_final_deaths_bedwars: number;
-  castle_entity_attack_final_kills_bedwars: number;
-  castle_entity_attack_kills_bedwars: number;
-  castle_entity_explosion_deaths_bedwars: number;
-  castle_entity_explosion_final_kills_bedwars: number;
-  castle_entity_explosion_kills_bedwars: number;
-  castle_fall_deaths_bedwars: number;
-  castle_fall_final_kills_bedwars: number;
-  castle_fall_kills_bedwars: number;
-  castle_final_deaths_bedwars: number;
-  castle_final_kills_bedwars: number;
-  castle_games_played_bedwars: number;
-  castle_gold_resources_collected_bedwars: number;
-  castle_iron_resources_collected_bedwars: number;
-  castle_items_purchased_bedwars: number;
-  castle_kills_bedwars: number;
-  castle_losses_bedwars: number;
-  'castle_permanent _items_purchased_bedwars': number;
-  castle_permanent_items_purchased_bedwars: number;
-  castle_resources_collected_bedwars: number;
-  castle_void_deaths_bedwars: number;
-  castle_void_final_kills_bedwars: number;
-  castle_void_kills_bedwars: number;
-  castle_wins_bedwars: number;
-  castle_winstreak: number;
-  coins: number;
-  deaths_bedwars: number;
-  diamond_resources_collected_bedwars: number;
-  eight_one__items_purchased_bedwars: number;
-  eight_one_beds_broken_bedwars: number;
-  eight_one_beds_lost_bedwars: number;
-  eight_one_deaths_bedwars: number;
-  eight_one_diamond_resources_collected_bedwars: number;
-  eight_one_emerald_resources_collected_bedwars: number;
-  eight_one_entity_attack_deaths_bedwars: number;
-  eight_one_entity_attack_final_deaths_bedwars: number;
-  eight_one_entity_attack_final_kills_bedwars: number;
-  eight_one_entity_attack_kills_bedwars: number;
-  eight_one_entity_explosion_deaths_bedwars: number;
-  eight_one_entity_explosion_final_deaths_bedwars: number;
-  eight_one_entity_explosion_final_kills_bedwars: number;
-  eight_one_fall_deaths_bedwars: number;
-  eight_one_fall_final_deaths_bedwars: number;
-  eight_one_fall_final_kills_bedwars: number;
-  eight_one_fall_kills_bedwars: number;
-  eight_one_final_deaths_bedwars: number;
-  eight_one_final_kills_bedwars: number;
-  eight_one_games_played_bedwars: number;
-  eight_one_gold_resources_collected_bedwars: number;
-  eight_one_iron_resources_collected_bedwars: number;
-  eight_one_items_purchased_bedwars: number;
-  eight_one_kills_bedwars: number;
-  eight_one_losses_bedwars: number;
-  eight_one_magic_deaths_bedwars: number;
-  eight_one_magic_final_deaths_bedwars: number;
-  eight_one_magic_final_kills_bedwars: number;
-  'eight_one_permanent _items_purchased_bedwars': number;
-  eight_one_permanent_items_purchased_bedwars: number;
-  eight_one_resources_collected_bedwars: number;
-  eight_one_suffocation_deaths_bedwars: number;
-  eight_one_void_deaths_bedwars: number;
-  eight_one_void_final_deaths_bedwars: number;
-  eight_one_void_final_kills_bedwars: number;
-  eight_one_void_kills_bedwars: number;
-  eight_one_wins_bedwars: number;
-  eight_one_winstreak: number;
-  eight_two__items_purchased_bedwars: number;
-  eight_two_armed__items_purchased_bedwars: number;
-  eight_two_armed_beds_broken_bedwars: number;
-  eight_two_armed_beds_lost_bedwars: number;
-  eight_two_armed_deaths_bedwars: number;
-  eight_two_armed_diamond_resources_collected_bedwars: number;
-  eight_two_armed_emerald_resources_collected_bedwars: number;
-  eight_two_armed_entity_attack_deaths_bedwars: number;
-  eight_two_armed_entity_attack_final_deaths_bedwars: number;
-  eight_two_armed_entity_attack_kills_bedwars: number;
-  eight_two_armed_entity_explosion_deaths_bedwars: number;
-  eight_two_armed_entity_explosion_final_kills_bedwars: number;
-  eight_two_armed_fall_deaths_bedwars: number;
-  eight_two_armed_fall_final_deaths_bedwars: number;
-  eight_two_armed_final_deaths_bedwars: number;
-  eight_two_armed_final_kills_bedwars: number;
-  eight_two_armed_games_played_bedwars: number;
-  eight_two_armed_gold_resources_collected_bedwars: number;
-  eight_two_armed_iron_resources_collected_bedwars: number;
-  eight_two_armed_items_purchased_bedwars: number;
-  eight_two_armed_kills_bedwars: number;
-  eight_two_armed_losses_bedwars: number;
-  eight_two_armed_magic_final_deaths_bedwars: number;
-  'eight_two_armed_permanent _items_purchased_bedwars': number;
-  eight_two_armed_projectile_deaths_bedwars: number;
-  eight_two_armed_projectile_final_deaths_bedwars: number;
-  eight_two_armed_projectile_final_kills_bedwars: number;
-  eight_two_armed_projectile_kills_bedwars: number;
-  eight_two_armed_resources_collected_bedwars: number;
-  eight_two_armed_void_deaths_bedwars: number;
-  eight_two_armed_void_final_kills_bedwars: number;
-  eight_two_armed_void_kills_bedwars: number;
-  eight_two_armed_winstreak: number;
-  eight_two_beds_broken_bedwars: number;
-  eight_two_beds_lost_bedwars: number;
-  eight_two_deaths_bedwars: number;
-  eight_two_diamond_resources_collected_bedwars: number;
-  eight_two_emerald_resources_collected_bedwars: number;
-  eight_two_entity_attack_deaths_bedwars: number;
-  eight_two_entity_attack_final_deaths_bedwars: number;
-  eight_two_entity_attack_final_kills_bedwars: number;
-  eight_two_entity_attack_kills_bedwars: number;
-  eight_two_entity_explosion_deaths_bedwars: number;
-  eight_two_entity_explosion_final_deaths_bedwars: number;
-  eight_two_entity_explosion_final_kills_bedwars: number;
-  eight_two_entity_explosion_kills_bedwars: number;
-  eight_two_fall_deaths_bedwars: number;
-  eight_two_fall_final_deaths_bedwars: number;
-  eight_two_fall_final_kills_bedwars: number;
-  eight_two_fall_kills_bedwars: number;
-  eight_two_final_deaths_bedwars: number;
-  eight_two_final_kills_bedwars: number;
-  eight_two_games_played_bedwars: number;
-  eight_two_gold_resources_collected_bedwars: number;
-  eight_two_iron_resources_collected_bedwars: number;
-  eight_two_items_purchased_bedwars: number;
-  eight_two_kills_bedwars: number;
-  eight_two_losses_bedwars: number;
-  eight_two_lucky__items_purchased_bedwars: number;
-  eight_two_lucky_beds_broken_bedwars: number;
-  eight_two_lucky_beds_lost_bedwars: number;
-  eight_two_lucky_deaths_bedwars: number;
-  eight_two_lucky_diamond_resources_collected_bedwars: number;
-  eight_two_lucky_emerald_resources_collected_bedwars: number;
-  eight_two_lucky_entity_attack_deaths_bedwars: number;
-  eight_two_lucky_entity_attack_final_deaths_bedwars: number;
-  eight_two_lucky_entity_attack_final_kills_bedwars: number;
-  eight_two_lucky_entity_attack_kills_bedwars: number;
-  eight_two_lucky_entity_explosion_final_kills_bedwars: number;
-  eight_two_lucky_final_deaths_bedwars: number;
-  eight_two_lucky_final_kills_bedwars: number;
-  eight_two_lucky_fire_tick_deaths_bedwars: number;
-  eight_two_lucky_fire_tick_final_deaths_bedwars: number;
-  eight_two_lucky_fire_tick_final_kills_bedwars: number;
-  eight_two_lucky_fire_tick_kills_bedwars: number;
-  eight_two_lucky_games_played_bedwars: number;
-  eight_two_lucky_gold_resources_collected_bedwars: number;
-  eight_two_lucky_iron_resources_collected_bedwars: number;
-  eight_two_lucky_items_purchased_bedwars: number;
-  eight_two_lucky_kills_bedwars: number;
-  eight_two_lucky_lava_kills_bedwars: number;
-  eight_two_lucky_losses_bedwars: number;
-  eight_two_lucky_magic_final_deaths_bedwars: number;
-  'eight_two_lucky_permanent _items_purchased_bedwars': number;
-  eight_two_lucky_resources_collected_bedwars: number;
-  eight_two_lucky_void_deaths_bedwars: number;
-  eight_two_lucky_void_final_deaths_bedwars: number;
-  eight_two_lucky_void_final_kills_bedwars: number;
-  eight_two_lucky_void_kills_bedwars: number;
-  eight_two_lucky_wins_bedwars: number;
-  eight_two_lucky_winstreak: number;
-  eight_two_magic_deaths_bedwars: number;
-  eight_two_magic_final_deaths_bedwars: number;
-  eight_two_magic_final_kills_bedwars: number;
-  eight_two_magic_kills_bedwars: number;
-  'eight_two_permanent _items_purchased_bedwars': number;
-  eight_two_permanent_items_purchased_bedwars: number;
-  eight_two_resources_collected_bedwars: number;
-  eight_two_rush__items_purchased_bedwars: number;
-  eight_two_rush_beds_broken_bedwars: number;
-  eight_two_rush_beds_lost_bedwars: number;
-  eight_two_rush_deaths_bedwars: number;
-  eight_two_rush_diamond_resources_collected_bedwars: number;
-  eight_two_rush_emerald_resources_collected_bedwars: number;
-  eight_two_rush_entity_attack_deaths_bedwars: number;
-  eight_two_rush_entity_attack_final_deaths_bedwars: number;
-  eight_two_rush_entity_attack_final_kills_bedwars: number;
-  eight_two_rush_entity_attack_kills_bedwars: number;
-  eight_two_rush_final_deaths_bedwars: number;
-  eight_two_rush_final_kills_bedwars: number;
-  eight_two_rush_games_played_bedwars: number;
-  eight_two_rush_gold_resources_collected_bedwars: number;
-  eight_two_rush_iron_resources_collected_bedwars: number;
-  eight_two_rush_items_purchased_bedwars: number;
-  eight_two_rush_kills_bedwars: number;
-  eight_two_rush_losses_bedwars: number;
-  'eight_two_rush_permanent _items_purchased_bedwars': number;
-  eight_two_rush_resources_collected_bedwars: number;
-  eight_two_rush_void_deaths_bedwars: number;
-  eight_two_rush_void_final_deaths_bedwars: number;
-  eight_two_rush_void_final_kills_bedwars: number;
-  eight_two_rush_void_kills_bedwars: number;
-  eight_two_rush_wins_bedwars: number;
-  eight_two_rush_winstreak: number;
-  eight_two_ultimate__items_purchased_bedwars: number;
-  eight_two_ultimate_beds_lost_bedwars: number;
-  eight_two_ultimate_deaths_bedwars: number;
-  eight_two_ultimate_entity_attack_deaths_bedwars: number;
-  eight_two_ultimate_entity_attack_final_deaths_bedwars: number;
-  eight_two_ultimate_entity_attack_kills_bedwars: number;
-  eight_two_ultimate_final_deaths_bedwars: number;
-  eight_two_ultimate_final_kills_bedwars: number;
-  eight_two_ultimate_games_played_bedwars: number;
-  eight_two_ultimate_gold_resources_collected_bedwars: number;
-  eight_two_ultimate_iron_resources_collected_bedwars: number;
-  eight_two_ultimate_items_purchased_bedwars: number;
-  eight_two_ultimate_kills_bedwars: number;
-  eight_two_ultimate_losses_bedwars: number;
-  eight_two_ultimate_permanent_items_purchased_bedwars: number;
-  eight_two_ultimate_resources_collected_bedwars: number;
-  eight_two_ultimate_void_deaths_bedwars: number;
-  eight_two_ultimate_void_final_deaths_bedwars: number;
-  eight_two_ultimate_void_final_kills_bedwars: number;
-  eight_two_ultimate_winstreak: number;
-  eight_two_void_deaths_bedwars: number;
-  eight_two_void_final_deaths_bedwars: number;
-  eight_two_void_final_kills_bedwars: number;
-  eight_two_void_kills_bedwars: number;
-  eight_two_voidless__items_purchased_bedwars: number;
-  eight_two_voidless_beds_broken_bedwars: number;
-  eight_two_voidless_beds_lost_bedwars: number;
-  eight_two_voidless_deaths_bedwars: number;
-  eight_two_voidless_diamond_resources_collected_bedwars: number;
-  eight_two_voidless_emerald_resources_collected_bedwars: number;
-  eight_two_voidless_entity_attack_deaths_bedwars: number;
-  eight_two_voidless_entity_attack_final_deaths_bedwars: number;
-  eight_two_voidless_entity_attack_final_kills_bedwars: number;
-  eight_two_voidless_entity_attack_kills_bedwars: number;
-  eight_two_voidless_entity_explosion_kills_bedwars: number;
-  eight_two_voidless_fall_deaths_bedwars: number;
-  eight_two_voidless_fall_kills_bedwars: number;
-  eight_two_voidless_final_deaths_bedwars: number;
-  eight_two_voidless_final_kills_bedwars: number;
-  eight_two_voidless_games_played_bedwars: number;
-  eight_two_voidless_gold_resources_collected_bedwars: number;
-  eight_two_voidless_iron_resources_collected_bedwars: number;
-  eight_two_voidless_items_purchased_bedwars: number;
-  eight_two_voidless_kills_bedwars: number;
-  eight_two_voidless_losses_bedwars: number;
-  'eight_two_voidless_permanent _items_purchased_bedwars': number;
-  eight_two_voidless_resources_collected_bedwars: number;
-  eight_two_voidless_wins_bedwars: number;
-  eight_two_voidless_winstreak: number;
-  eight_two_wins_bedwars: number;
-  eight_two_winstreak: number;
-  emerald_resources_collected_bedwars: number;
-  entity_attack_deaths_bedwars: number;
-  entity_attack_final_deaths_bedwars: number;
-  entity_attack_final_kills_bedwars: number;
-  entity_attack_kills_bedwars: number;
-  entity_explosion_deaths_bedwars: number;
-  entity_explosion_final_deaths_bedwars: number;
-  entity_explosion_final_kills_bedwars: number;
-  entity_explosion_kills_bedwars: number;
-  Experience: number;
-  fall_deaths_bedwars: number;
-  fall_final_deaths_bedwars: number;
-  fall_final_kills_bedwars: number;
-  fall_kills_bedwars: number;
-  final_deaths_bedwars: number;
-  final_kills_bedwars: number;
-  fire_deaths_bedwars: number;
-  fire_tick_deaths_bedwars: number;
-  first_join_7: true;
-  four_four__items_purchased_bedwars: number;
-  four_four_armed__items_purchased_bedwars: number;
-  four_four_armed_beds_lost_bedwars: number;
-  four_four_armed_deaths_bedwars: number;
-  four_four_armed_diamond_resources_collected_bedwars: number;
-  four_four_armed_emerald_resources_collected_bedwars: number;
-  four_four_armed_entity_attack_deaths_bedwars: number;
-  four_four_armed_entity_attack_final_kills_bedwars: number;
-  four_four_armed_entity_attack_kills_bedwars: number;
-  four_four_armed_fall_deaths_bedwars: number;
-  four_four_armed_final_deaths_bedwars: number;
-  four_four_armed_final_kills_bedwars: number;
-  four_four_armed_games_played_bedwars: number;
-  four_four_armed_gold_resources_collected_bedwars: number;
-  four_four_armed_iron_resources_collected_bedwars: number;
-  four_four_armed_items_purchased_bedwars: number;
-  four_four_armed_kills_bedwars: number;
-  four_four_armed_losses_bedwars: number;
-  'four_four_armed_permanent _items_purchased_bedwars': number;
-  four_four_armed_permanent_items_purchased_bedwars: number;
-  four_four_armed_projectile_deaths_bedwars: number;
-  four_four_armed_projectile_final_deaths_bedwars: number;
-  four_four_armed_projectile_kills_bedwars: number;
-  four_four_armed_resources_collected_bedwars: number;
-  four_four_armed_void_deaths_bedwars: number;
-  four_four_armed_void_kills_bedwars: number;
-  four_four_armed_wins_bedwars: number;
-  four_four_armed_winstreak: number;
-  four_four_beds_broken_bedwars: number;
-  four_four_beds_lost_bedwars: number;
-  four_four_deaths_bedwars: number;
-  four_four_diamond_resources_collected_bedwars: number;
-  four_four_emerald_resources_collected_bedwars: number;
-  four_four_entity_attack_deaths_bedwars: number;
-  four_four_entity_attack_final_deaths_bedwars: number;
-  four_four_entity_attack_final_kills_bedwars: number;
-  four_four_entity_attack_kills_bedwars: number;
-  four_four_entity_explosion_deaths_bedwars: number;
-  four_four_entity_explosion_final_kills_bedwars: number;
-  four_four_entity_explosion_kills_bedwars: number;
-  four_four_fall_deaths_bedwars: number;
-  four_four_fall_final_deaths_bedwars: number;
-  four_four_fall_final_kills_bedwars: number;
-  four_four_fall_kills_bedwars: number;
-  four_four_final_deaths_bedwars: number;
-  four_four_final_kills_bedwars: number;
-  four_four_games_played_bedwars: number;
-  four_four_gold_resources_collected_bedwars: number;
-  four_four_iron_resources_collected_bedwars: number;
-  four_four_items_purchased_bedwars: number;
-  four_four_kills_bedwars: number;
-  four_four_losses_bedwars: number;
-  four_four_lucky__items_purchased_bedwars: number;
-  four_four_lucky_beds_broken_bedwars: number;
-  four_four_lucky_beds_lost_bedwars: number;
-  four_four_lucky_deaths_bedwars: number;
-  four_four_lucky_diamond_resources_collected_bedwars: number;
-  four_four_lucky_emerald_resources_collected_bedwars: number;
-  four_four_lucky_entity_attack_deaths_bedwars: number;
-  four_four_lucky_entity_attack_final_deaths_bedwars: number;
-  four_four_lucky_entity_attack_final_kills_bedwars: number;
-  four_four_lucky_entity_attack_kills_bedwars: number;
-  four_four_lucky_final_deaths_bedwars: number;
-  four_four_lucky_final_kills_bedwars: number;
-  four_four_lucky_games_played_bedwars: number;
-  four_four_lucky_gold_resources_collected_bedwars: number;
-  four_four_lucky_iron_resources_collected_bedwars: number;
-  four_four_lucky_items_purchased_bedwars: number;
-  four_four_lucky_kills_bedwars: number;
-  four_four_lucky_losses_bedwars: number;
-  four_four_lucky_magic_final_deaths_bedwars: number;
-  four_four_lucky_magic_kills_bedwars: number;
-  'four_four_lucky_permanent _items_purchased_bedwars': number;
-  four_four_lucky_permanent_items_purchased_bedwars: number;
-  four_four_lucky_resources_collected_bedwars: number;
-  four_four_lucky_void_deaths_bedwars: number;
-  four_four_lucky_void_final_deaths_bedwars: number;
-  four_four_lucky_void_final_kills_bedwars: number;
-  four_four_lucky_void_kills_bedwars: number;
-  four_four_lucky_wins_bedwars: number;
-  four_four_lucky_winstreak: number;
-  four_four_magic_final_deaths_bedwars: number;
-  four_four_magic_final_kills_bedwars: number;
-  four_four_magic_kills_bedwars: number;
-  'four_four_permanent _items_purchased_bedwars': number;
-  four_four_permanent_items_purchased_bedwars: number;
-  four_four_resources_collected_bedwars: number;
-  four_four_rush__items_purchased_bedwars: number;
-  four_four_rush_beds_broken_bedwars: number;
-  four_four_rush_beds_lost_bedwars: number;
-  four_four_rush_deaths_bedwars: number;
-  four_four_rush_diamond_resources_collected_bedwars: number;
-  four_four_rush_emerald_resources_collected_bedwars: number;
-  four_four_rush_entity_attack_deaths_bedwars: number;
-  four_four_rush_entity_attack_final_deaths_bedwars: number;
-  four_four_rush_entity_attack_final_kills_bedwars: number;
-  four_four_rush_entity_attack_kills_bedwars: number;
-  four_four_rush_final_deaths_bedwars: number;
-  four_four_rush_final_kills_bedwars: number;
-  four_four_rush_games_played_bedwars: number;
-  four_four_rush_gold_resources_collected_bedwars: number;
-  four_four_rush_iron_resources_collected_bedwars: number;
-  four_four_rush_items_purchased_bedwars: number;
-  four_four_rush_kills_bedwars: number;
-  four_four_rush_losses_bedwars: number;
-  'four_four_rush_permanent _items_purchased_bedwars': number;
-  four_four_rush_permanent_items_purchased_bedwars: number;
-  four_four_rush_resources_collected_bedwars: number;
-  four_four_rush_suffocation_deaths_bedwars: number;
-  four_four_rush_void_deaths_bedwars: number;
-  four_four_rush_void_final_deaths_bedwars: number;
-  four_four_rush_void_final_kills_bedwars: number;
-  four_four_rush_void_kills_bedwars: number;
-  four_four_rush_wins_bedwars: number;
-  four_four_rush_winstreak: number;
-  four_four_ultimate__items_purchased_bedwars: number;
-  four_four_ultimate_beds_broken_bedwars: number;
-  four_four_ultimate_beds_lost_bedwars: number;
-  four_four_ultimate_deaths_bedwars: number;
-  four_four_ultimate_diamond_resources_collected_bedwars: number;
-  four_four_ultimate_emerald_resources_collected_bedwars: number;
-  four_four_ultimate_entity_attack_deaths_bedwars: number;
-  four_four_ultimate_entity_attack_final_deaths_bedwars: number;
-  four_four_ultimate_entity_attack_final_kills_bedwars: number;
-  four_four_ultimate_entity_attack_kills_bedwars: number;
-  four_four_ultimate_final_deaths_bedwars: number;
-  four_four_ultimate_final_kills_bedwars: number;
-  four_four_ultimate_games_played_bedwars: number;
-  four_four_ultimate_gold_resources_collected_bedwars: number;
-  four_four_ultimate_iron_resources_collected_bedwars: number;
-  four_four_ultimate_items_purchased_bedwars: number;
-  four_four_ultimate_kills_bedwars: number;
-  four_four_ultimate_losses_bedwars: number;
-  'four_four_ultimate_permanent _items_purchased_bedwars': number;
-  four_four_ultimate_resources_collected_bedwars: number;
-  four_four_ultimate_void_deaths_bedwars: number;
-  four_four_ultimate_void_final_deaths_bedwars: number;
-  four_four_ultimate_void_final_kills_bedwars: number;
-  four_four_ultimate_void_kills_bedwars: number;
-  four_four_ultimate_wins_bedwars: number;
-  four_four_ultimate_winstreak: number;
-  four_four_void_deaths_bedwars: number;
-  four_four_void_final_deaths_bedwars: number;
-  four_four_void_final_kills_bedwars: number;
-  four_four_void_kills_bedwars: number;
-  four_four_voidless__items_purchased_bedwars: number;
-  four_four_voidless_beds_broken_bedwars: number;
-  four_four_voidless_beds_lost_bedwars: number;
-  four_four_voidless_deaths_bedwars: number;
-  four_four_voidless_diamond_resources_collected_bedwars: number;
-  four_four_voidless_entity_attack_deaths_bedwars: number;
-  four_four_voidless_entity_attack_final_deaths_bedwars: number;
-  four_four_voidless_entity_attack_final_kills_bedwars: number;
-  four_four_voidless_entity_attack_kills_bedwars: number;
-  four_four_voidless_entity_explosion_final_kills_bedwars: number;
-  four_four_voidless_fall_final_deaths_bedwars: number;
-  four_four_voidless_final_deaths_bedwars: number;
-  four_four_voidless_final_kills_bedwars: number;
-  four_four_voidless_games_played_bedwars: number;
-  four_four_voidless_gold_resources_collected_bedwars: number;
-  four_four_voidless_iron_resources_collected_bedwars: number;
-  four_four_voidless_items_purchased_bedwars: number;
-  four_four_voidless_kills_bedwars: number;
-  four_four_voidless_losses_bedwars: number;
-  'four_four_voidless_permanent _items_purchased_bedwars': number;
-  four_four_voidless_permanent_items_purchased_bedwars: number;
-  four_four_voidless_resources_collected_bedwars: number;
-  four_four_voidless_void_final_deaths_bedwars: number;
-  four_four_voidless_wins_bedwars: number;
-  four_four_voidless_winstreak: number;
-  four_four_wins_bedwars: number;
-  four_four_winstreak: number;
-  four_three__items_purchased_bedwars: number;
-  four_three_beds_broken_bedwars: number;
-  four_three_beds_lost_bedwars: number;
-  four_three_deaths_bedwars: number;
-  four_three_diamond_resources_collected_bedwars: number;
-  four_three_emerald_resources_collected_bedwars: number;
-  four_three_entity_attack_deaths_bedwars: number;
-  four_three_entity_attack_final_deaths_bedwars: number;
-  four_three_entity_attack_final_kills_bedwars: number;
-  four_three_entity_attack_kills_bedwars: number;
-  four_three_entity_explosion_deaths_bedwars: number;
-  four_three_entity_explosion_kills_bedwars: number;
-  four_three_fall_deaths_bedwars: number;
-  four_three_fall_final_deaths_bedwars: number;
-  four_three_fall_final_kills_bedwars: number;
-  four_three_fall_kills_bedwars: number;
-  four_three_final_deaths_bedwars: number;
-  four_three_final_kills_bedwars: number;
-  four_three_fire_deaths_bedwars: number;
-  four_three_fire_tick_deaths_bedwars: number;
-  four_three_games_played_bedwars: number;
-  four_three_gold_resources_collected_bedwars: number;
-  four_three_iron_resources_collected_bedwars: number;
-  four_three_items_purchased_bedwars: number;
-  four_three_kills_bedwars: number;
-  four_three_losses_bedwars: number;
-  four_three_magic_deaths_bedwars: number;
-  four_three_magic_final_deaths_bedwars: number;
-  four_three_magic_final_kills_bedwars: number;
-  four_three_magic_kills_bedwars: number;
-  'four_three_permanent _items_purchased_bedwars': number;
-  four_three_permanent_items_purchased_bedwars: number;
-  four_three_projectile_deaths_bedwars: number;
-  four_three_resources_collected_bedwars: number;
-  four_three_void_deaths_bedwars: number;
-  four_three_void_final_deaths_bedwars: number;
-  four_three_void_final_kills_bedwars: number;
-  four_three_void_kills_bedwars: number;
-  four_three_wins_bedwars: number;
-  four_three_winstreak: number;
-  games_played_bedwars_1: number;
-  games_played_bedwars: number;
-  gold_resources_collected_bedwars: number;
-  iron_resources_collected_bedwars: number;
-  items_purchased_bedwars: number;
-  kills_bedwars: number;
-  losses_bedwars: number;
-  magic_deaths_bedwars: number;
-  magic_final_deaths_bedwars: number;
-  magic_final_kills_bedwars: number;
-  magic_kills_bedwars: number;
-  'permanent _items_purchased_bedwars': number;
-  permanent_items_purchased_bedwars: number;
-  projectile_deaths_bedwars: number;
-  resources_collected_bedwars: number;
-  suffocation_deaths_bedwars: number;
-  two_four__items_purchased_bedwars: number;
-  two_four_beds_broken_bedwars: number;
-  two_four_beds_lost_bedwars: number;
-  two_four_deaths_bedwars: number;
-  two_four_entity_attack_deaths_bedwars: number;
-  two_four_entity_attack_final_deaths_bedwars: number;
-  two_four_entity_attack_final_kills_bedwars: number;
-  two_four_entity_attack_kills_bedwars: number;
-  two_four_fall_deaths_bedwars: number;
-  two_four_final_deaths_bedwars: number;
-  two_four_final_kills_bedwars: number;
-  two_four_games_played_bedwars: number;
-  two_four_gold_resources_collected_bedwars: number;
-  two_four_iron_resources_collected_bedwars: number;
-  two_four_items_purchased_bedwars: number;
-  two_four_kills_bedwars: number;
-  two_four_losses_bedwars: number;
-  'two_four_permanent _items_purchased_bedwars': number;
-  two_four_permanent_items_purchased_bedwars: number;
-  two_four_resources_collected_bedwars: number;
-  two_four_void_deaths_bedwars: number;
-  two_four_void_final_deaths_bedwars: number;
-  two_four_void_kills_bedwars: number;
-  two_four_wins_bedwars: number;
-  two_four_winstreak: number;
-  void_deaths_bedwars: number;
-  void_final_deaths_bedwars: number;
-  void_final_kills_bedwars: number;
-  void_kills_bedwars: number;
-  wins_bedwars: number;
-  winstreak: number;
-}
