+++ src/listeners/voice.ts
@@ -1,13 +1,13 @@
 import { VoiceState } from 'discord.js';
 
-import { client, queueStore } from '..';
+import { client, getLogger, LoggerType, queueStore } from '../providers';
 
 export const onVoiceStateUpdate = () => (oldState: VoiceState, newState: VoiceState): void => {
   if (oldState.id === client.user?.id && newState.id === client.user?.id) {
     // that's us!
     if (oldState.channelID != null && newState.channelID == null) {
       // disconnected
-      console.log('bot was disconnected');
+      getLogger(LoggerType.VOICE, newState.id).info('bot was disconnected in ' + newState.guild.id);
       const queue = queueStore.get(newState.guild.id);
       if (!queue) return;
       delete queue.voiceChannel;
