+++ src/util/index.ts
@@ -1,7 +1,10 @@
 export * from './db';
 export * from './duration';
+export * from './embed';
 export * from './flags';
 export * from './map';
 export * from './message';
+export * from './music';
 export * from './path';
 export * from './regexp';
+export * from './store';
