+++ src/util/store.ts
@@ -3,7 +3,7 @@ import fse from 'fs-extra';
 import yaml from 'js-yaml';
 import _ from 'lodash/fp';
 
-import { mapToObject, objectToMap } from '.';
+import { mapToObject, objectToMap } from '../util';
 
 export interface StoreOptions {
   /** Path to file to store data in. */
